+++ src/com/intellij/idea/plugin/hybris/system/bean/codeInsight/completion/BSCompletionContributor.kt
@@ -22,6 +22,7 @@ import com.intellij.codeInsight.completion.CompletionContributor
 import com.intellij.codeInsight.completion.CompletionType
 import com.intellij.idea.plugin.hybris.system.bean.codeInsight.completion.provider.BSBeanClassCompletionProvider
 import com.intellij.idea.plugin.hybris.system.bean.codeInsight.completion.provider.BSBeanPropertyTypeCompletionProvider
+import com.intellij.idea.plugin.hybris.system.bean.codeInsight.completion.provider.BSEnumClassCompletionProvider
 import com.intellij.idea.plugin.hybris.system.bean.psi.BSPatterns
 import com.intellij.patterns.PlatformPatterns
 
@@ -32,6 +33,11 @@ class BSCompletionContributor : CompletionContributor() {
             CompletionType.BASIC,
             PlatformPatterns.psiElement().inside(BSPatterns.BEAN_CLASS),
             BSBeanClassCompletionProvider.instance
+        )
+        extend(
+            CompletionType.BASIC,
+            PlatformPatterns.psiElement().inside(BSPatterns.ENUM_CLASS),
+            BSEnumClassCompletionProvider.instance
         )
         extend(
             CompletionType.BASIC,